Allow aliasing operators and functions with different return type
Currently, the return type of an EdgeQL function or operator must match
that of the SQL function or operator it aliases.  It is possible to
workaround by using a `FROM SQL` wrapper, but that is clunky and may
affect performance due to an extra function indirection.  Aggregates do
not have even this workaround.

To fix this, introduce a new std-only `force_return_cast` field.  When
set, it instructs the compiler to automatically insert a cast into the
declared return type.  Convert some std operators and functions which
were relying on wrappers, and fix the return type of certain math
aggregates.

+++ tests/test_edgeql_ir_scopetree.py
@@ -448,7 +448,7 @@ class TestEdgeQLIRScopeTree(tb.BaseEdgeQLCompilerTest):
                         "(test::User)"
                     },
                     "(test::User).>(test::friends)[IS test::User]\
-.>(test::deck_cost)[IS std::decimal]": {
+.>(test::deck_cost)[IS std::int64]": {
                         "FENCE": {
                             "FENCE": {
                                 "FENCE": {
@@ -469,7 +469,7 @@ class TestEdgeQLIRScopeTree(tb.BaseEdgeQLCompilerTest):
             },
             "FENCE": {
                 "(__derived__::__derived__|x@@w~1)\
-.>(__tuple__::0)[IS std::decimal]"
+.>(__tuple__::0)[IS std::float64]"
             }
         }
         """
